separate type defs
